+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+### Unreleased
+
+##### Changed
+
+* Change module name from `platanus.asyncUpload` to `platanus.pallet`
+* Change directive name from `async-upload` to `pallet-file-selector`
+
+### v1.1.0
+
+##### Changed
+
+* Include scss and not uglyfied source in dist directory
+
 ### v1.0.0
 
 * Initial release.